Fix for udAllocStack
git-svn-id: https://bne-fs-pf-001.euclideon.local/svn/Dev/branch/pev/iterative_test/src@597 f54097bf-93f5-7042-b87d-fd266665e7ff

+++ udPlatform/Source/udFileHandler_HTTP.cpp
@@ -256,7 +256,7 @@ udResult udFileHandler_HTTPOpen(udFile **ppFile, const char *pFilename, udFileOp
   udFile_HTTP *pFile = nullptr;
   struct hostent *hp;
   
-  size_t actualHeaderLen;
+  int actualHeaderLen;
 
   // Automatically fail if trying to write to files on http
   result = udR_File_OpenFailure;
